+++ src/net/eathena/chathandler.cpp
@@ -757,7 +757,7 @@ void ChatHandler::talkPet(const std::string &restrict text,
 
     createOutPacket(CMSG_PET_TALK);
     outMsg.writeInt16(static_cast<int16_t>(sz + 4 + 1), "len");
-    outMsg.writeString(msg, sz, "message");
+    outMsg.writeString(msg, static_cast<int>(sz), "message");
     outMsg.writeInt8(0, "zero byte");
 }
